refactor: reorganize server managers into protocol domains
What changed and why:
- Moved protocol domain managers (tools, resources, prompts, logging,
  completions) to server/protocol/
- Kept infrastructure components (client_manager, coordinator) and user
  event system (callbacks) at server level
- Updated test structure to mirror the new source organization
- Moved protocol domain tests to tests/server/protocol/

Impact:
- Clear architectural distinction between protocol domains and
  infrastructure
- Reduced cognitive load when navigating the codebase
- Protocol domain responsibilities are now explicitly grouped
- Test organization matches source structure for easier navigation
- Infrastructure components are clearly identified at the top level
