+++ src/conduit/server/protocol/completions.py
