+++ src/conduit/server/session.py
@@ -55,15 +55,15 @@ from conduit.protocol.tools import (
     ListToolsRequest,
     ListToolsResult,
 )
-from conduit.server.managers.callbacks import CallbackManager
-from conduit.server.managers.completions import (
+from conduit.server.callbacks import CallbackManager
+from conduit.server.protocol.completions import (
     CompletionManager,
     CompletionNotConfiguredError,
 )
-from conduit.server.managers.logging import LoggingManager
-from conduit.server.managers.prompts import PromptManager
-from conduit.server.managers.resources import ResourceManager
-from conduit.server.managers.tools import ToolManager
+from conduit.server.protocol.logging import LoggingManager
+from conduit.server.protocol.prompts import PromptManager
+from conduit.server.protocol.resources import ResourceManager
+from conduit.server.protocol.tools import ToolManager
 from conduit.shared.session import BaseSession
 from conduit.transport.base import Transport
